added scroll detector
