+++ src/components/functions/scroll.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+const scroll = (setShowTopBar) => {
+  const handleScroll = () => {
+    const position = window.scrollY;
+    if (position >= 2) {
+      return setShowTopBar(true);
+    }
+    return setShowTopBar(false);
+  };
+  handleScroll();
+  window.addEventListener("scroll", handleScroll);
+  return () => {
+    window.removeEventListener("scroll", handleScroll);
+  };
+};
+
+export default scroll;
